completed initial work on QR code scanning
After getting CameraX set up, I hooked it up to MlKit and just made it
print out any QR Codes it scans. For now though I haven't tested it
with real QR codes as it took a while to get everything to build (I had
a strange error with Kotlin? Even though we're not using Kotlin?), but
everything seems to work now.