Fixed bug where score would be incorrect upon deleting a QR code
The main problem here was that we were storing the user's total score
on Firebase to begin with. This is something that can be calculated
given the list of hashes, and its easier to forget to update the total
score instead of just calling the method that gets the total score.