Fixed bug where deleted QR codes' comments would remain on the profile screen
This bug ended up being a lot deeper than I initially thought.
Essentially the problem was both a duplication of state locally on the
app, and an inconsistency with how different parts of the codebase
assume the database looks like.

What I did was consolidate all the local state into methods in the User
class, and ensure that code only runs when we know the database is in a
state we expect.